Main

+++ baranitech-be/database/migrations/2025_05_06_025737_create_trainings_table.php
@@ -10,23 +10,25 @@ return new class extends Migration {
      */
     public function up(): void
     {
-        Schema::create('trainings', function (Blueprint $table) {
-            $table->id();
-            $table->string('title');
-            $table->text('description')->nullable();
-            $table->foreignId('type')->constrained('training_categories')->onDelete('cascade');
-            $table->tinyInteger('classification'); // 0 => direct, 1 => online
-            $table->date('startdate');
-            $table->date('enddate')->nullable();
-            $table->string('location')->nullable();
-            $table->decimal('total_hours', 5, 2)->nullable();
-            $table->string('city')->nullable();
-            $table->string('state')->nullable();
-            $table->string('country')->nullable();
-            $table->text('table_of_contents')->nullable();
-            $table->decimal('total_price', 10, 2)->nullable();
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('trainings')) {
+            Schema::create('trainings', function (Blueprint $table) {
+                $table->id();
+                $table->string('title');
+                $table->text('description')->nullable();
+                $table->foreignId('type')->constrained('training_categories')->onDelete('cascade');
+                $table->tinyInteger('classification'); // 0 => direct, 1 => online
+                $table->date('startdate');
+                $table->date('enddate')->nullable();
+                $table->string('location')->nullable();
+                $table->decimal('total_hours', 5, 2)->nullable();
+                $table->string('city')->nullable();
+                $table->string('state')->nullable();
+                $table->string('country')->nullable();
+                $table->text('table_of_contents')->nullable();
+                $table->decimal('total_price', 10, 2)->nullable();
+                $table->timestamps();
+            });
+        }
     }
 
     /**
